requirements.txt
gensim==3.8.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
matplotlib==3.1.1
multidict==4.5.2
nltk==3.4.4
numpy==1.16.4
pandas==0.24.2
python-dateutil==2.8.0
python-twitter==3.5
scikit-learn==0.21.2
scipy==1.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tf-estimator-nightly==1.14.0.dev2019060501
-e git+https://github.com/twintproject/twint.git@cb28f2e1d53e0422dcf360c2a37cbb5defd5e0b7#egg=twint